Added base test module
